working permission and user location (poc)

+++ app/src/main/java/com/example/villomap/data/Geometry.kt
@@ -3,5 +3,5 @@ package com.example.villomap.data
 data class Geometry(
     val type: String,
     // coordinates = [longitude, latitude]
-    val coordinates: List<Float>,
+    val coordinates: List<Double>,
 )
